+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Added `month` and `tags` input plugins.
 Added `input` jQuery method.
 Added `readonly`, `disabled` and `value` features.
 Redesigned framework.
+Renamed project.
 
 # v0.1.1 / 2016-07-28