+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# v0.2.1 / 2016-08-15
+
+Added `file` and `autocomplete` input plugins.
+
 # v0.2.0 / 2016-08-01
 
 Added `month` and `tags` input plugins.